Refactor common-ast
