+++ common/ast/Cargo.toml
@@ -31,7 +31,6 @@ logos = "0.12.1"
 nom = "7.1.1"
 nom-rule = "0.3.0"
 pratt = "0.3.0"
-serde = { version = "1.0.137", features = ["derive"] }
 thiserror = "1.0.31"
 url = "2.2.2"
