+++ common/ast/src/ast/statements/kill.rs
@@ -1,0 +1,31 @@
+// Copyright 2022 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::fmt::Display;
+use std::fmt::Formatter;
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum KillTarget {
+    Query,
+    Connection,
+}
+
+impl Display for KillTarget {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            KillTarget::Query => write!(f, "QUERY"),
+            KillTarget::Connection => write!(f, "CONNECTION"),
+        }
+    }
+}
