+++ common/ast/src/ast/statements/mod.rs
@@ -1,0 +1,37 @@
+// Copyright 2021 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+mod copy;
+mod database;
+mod explain;
+mod insert;
+mod kill;
+mod show;
+mod stage;
+mod statement;
+mod table;
+mod user;
+mod view;
+
+pub use copy::*;
+pub use database::*;
+pub use explain::*;
+pub use insert::*;
+pub use kill::*;
+pub use show::*;
+pub use stage::*;
+pub use statement::*;
+pub use table::*;
+pub use user::*;
+pub use view::*;
