+++ common/ast/src/ast/statements/statement.rs
@@ -1,4 +1,4 @@
-// Copyright 2021 Datafuse Labs.
+// Copyright 2022 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,30 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::collections::BTreeMap;
 use std::fmt::Display;
 use std::fmt::Formatter;
 
-use common_meta_types::AuthType;
 use common_meta_types::PrincipalIdentity;
 use common_meta_types::UserIdentity;
-use common_meta_types::UserOption;
-use common_meta_types::UserOptionFlag;
-use common_meta_types::UserPrivilegeType;
-use serde::Deserialize;
-use serde::Serialize;
 
-use super::write_space_seperated_list;
-use super::Expr;
-use crate::ast::expr::Literal;
-use crate::ast::expr::TypeName;
+use super::*;
 use crate::ast::write_comma_separated_list;
 use crate::ast::write_period_separated_list;
 use crate::ast::write_quoted_comma_separated_list;
-use crate::ast::write_space_seperated_map;
+use crate::ast::write_space_seperated_list;
+use crate::ast::Expr;
 use crate::ast::Identifier;
+use crate::ast::Literal;
 use crate::ast::Query;
-use crate::parser::token::Token;
 
 // SQL statement
 #[derive(Debug, Clone, PartialEq)]
@@ -156,648 +147,6 @@ pub enum Statement<'a> {
         location: String,
         pattern: String,
     },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum ExplainKind {
-    Syntax,
-    Graph,
-    Pipeline,
-}
-#[derive(Debug, Clone, PartialEq)]
-pub struct InsertStmt<'a> {
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub table: Identifier<'a>,
-    pub columns: Vec<Identifier<'a>>,
-    pub source: InsertSource<'a>,
-    pub overwrite: bool,
-}
-
-/// CopyStmt is the parsed statement of `COPY`.
-///
-/// ## Examples
-///
-/// ```sql
-/// COPY INTO table from s3://bucket/path/to/x.csv
-/// ```
-#[derive(Debug, Clone, PartialEq)]
-pub struct CopyStmt<'a> {
-    pub src: CopyUnit<'a>,
-    pub dst: CopyUnit<'a>,
-    pub files: Vec<String>,
-    pub pattern: String,
-    pub file_format: BTreeMap<String, String>,
-    /// TODO(xuanwo): parse into validation_mode directly.
-    pub validation_mode: String,
-    pub size_limit: usize,
-}
-
-#[derive(Debug, Clone, PartialEq)] // Databases
-pub struct ShowDatabasesStmt<'a> {
-    pub limit: Option<ShowLimit<'a>>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ShowCreateDatabaseStmt<'a> {
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Identifier<'a>,
-}
-
-#[derive(Debug, Clone, PartialEq)] // Tables
-pub struct ShowTablesStmt<'a> {
-    pub database: Option<Identifier<'a>>,
-    pub full: bool,
-    pub limit: Option<ShowLimit<'a>>,
-    pub with_history: bool,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ShowCreateTableStmt<'a> {
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub table: Identifier<'a>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ShowTablesStatusStmt<'a> {
-    pub database: Option<Identifier<'a>>,
-    pub limit: Option<ShowLimit<'a>>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct CreateDatabaseStmt<'a> {
-    pub if_not_exists: bool,
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Identifier<'a>,
-    pub engine: Option<DatabaseEngine>,
-    pub options: Vec<SQLProperty>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct DropDatabaseStmt<'a> {
-    pub if_exists: bool,
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Identifier<'a>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct AlterDatabaseStmt<'a> {
-    pub if_exists: bool,
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Identifier<'a>,
-    pub action: AlterDatabaseAction<'a>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct CreateTableStmt<'a> {
-    pub if_not_exists: bool,
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub table: Identifier<'a>,
-    pub source: Option<CreateTableSource<'a>>,
-    pub table_options: Vec<TableOption>,
-    pub cluster_by: Vec<Expr<'a>>,
-    pub as_query: Option<Box<Query<'a>>>,
-    pub comment: Option<String>,
-    pub transient: bool,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct DescribeTableStmt<'a> {
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub table: Identifier<'a>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum CreateTableSource<'a> {
-    Columns(Vec<ColumnDefinition<'a>>),
-    Like {
-        catalog: Option<Identifier<'a>>,
-        database: Option<Identifier<'a>>,
-        table: Identifier<'a>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct DropTableStmt<'a> {
-    pub if_exists: bool,
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub table: Identifier<'a>,
-    pub all: bool,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct UndropTableStmt<'a> {
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub table: Identifier<'a>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct AlterTableStmt<'a> {
-    pub if_exists: bool,
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub table: Identifier<'a>,
-    pub action: AlterTableAction<'a>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct RenameTableStmt<'a> {
-    pub if_exists: bool,
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub table: Identifier<'a>,
-    pub new_catalog: Option<Identifier<'a>>,
-    pub new_database: Option<Identifier<'a>>,
-    pub new_table: Identifier<'a>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct TruncateTableStmt<'a> {
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub table: Identifier<'a>,
-    pub purge: bool,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct OptimizeTableStmt<'a> {
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub table: Identifier<'a>,
-    pub action: Option<OptimizeTableAction>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-pub enum OptimizeTableAction {
-    All,
-    Purge,
-    Compact,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ExistsTableStmt<'a> {
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub table: Identifier<'a>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum TableOption {
-    Engine(Engine),
-    Comment(String),
-}
-
-impl TableOption {
-    pub fn option_key(&self) -> String {
-        match self {
-            TableOption::Engine(_) => "ENGINE".to_string(),
-            TableOption::Comment(_) => "COMMENT".to_string(),
-        }
-    }
-
-    pub fn option_value(&self) -> String {
-        match self {
-            TableOption::Engine(engine) => engine.to_string(),
-            TableOption::Comment(comment) => comment.clone(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum Engine {
-    Null,
-    Memory,
-    Fuse,
-    Github,
-    View,
-    Random,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum DatabaseEngine {
-    Default,
-    Github(String),
-}
-
-/// CopyUnit is the unit that can be used in `COPY`.
-#[derive(Debug, Clone, PartialEq)]
-pub enum CopyUnit<'a> {
-    /// Table can be used in `INTO` or `FROM`.
-    ///
-    /// While table used as `FROM`, it will be rewrite as `(SELECT * FROM table)`
-    Table {
-        catalog: Option<Identifier<'a>>,
-        database: Option<Identifier<'a>>,
-        table: Identifier<'a>,
-    },
-    /// StageLocation (a.k.a internal and external stage) can be used
-    /// in `INTO` or `FROM`.
-    ///
-    /// For examples:
-    ///
-    /// - internal stage: `@internal_stage/path/to/dir/`
-    /// - external stage: `@s3_external_stage/path/to/dir/`
-    StageLocation {
-        /// The name of the stage.
-        name: String,
-        path: String,
-    },
-    /// UriLocation (a.k.a external location) can be used in `INTO` or `FROM`.
-    ///
-    /// For examples: `'s3://example/path/to/dir' CREDENTIALS = (AWS_ACCESS_ID="admin" AWS_SECRET_KEY="admin")`
-    ///
-    /// TODO(xuanwo): Add endpoint_url support.
-    /// TODO(xuanwo): We can check if we support this protocol during parsing.
-    /// TODO(xuanwo): Maybe we can introduce more strict (friendly) report for credentials and encryption, like parsed into StorageConfig?
-    UriLocation {
-        protocol: String,
-        name: String,
-        path: String,
-        credentials: BTreeMap<String, String>,
-        encryption: BTreeMap<String, String>,
-    },
-    /// Query can only be used as `FROM`.
-    ///
-    /// For example:`(SELECT field_a,field_b FROM table)`
-    Query(Box<Query<'a>>),
-}
-
-impl CopyUnit<'_> {
-    pub fn target(&self) -> &'static str {
-        match self {
-            CopyUnit::Table { .. } => "Table",
-            CopyUnit::StageLocation { .. } => "StageLocation",
-            CopyUnit::UriLocation { .. } => "UriLocation",
-            CopyUnit::Query(_) => "Query",
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct CreateViewStmt<'a> {
-    pub if_not_exists: bool,
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub view: Identifier<'a>,
-    pub query: Box<Query<'a>>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct AlterViewStmt<'a> {
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub view: Identifier<'a>,
-    pub query: Box<Query<'a>>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct DropViewStmt<'a> {
-    pub if_exists: bool,
-    pub catalog: Option<Identifier<'a>>,
-    pub database: Option<Identifier<'a>>,
-    pub view: Identifier<'a>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct SQLProperty {
-    pub name: String,
-    pub value: String,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum ShowLimit<'a> {
-    Like { pattern: String },
-    Where { selection: Box<Expr<'a>> },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ColumnDefinition<'a> {
-    pub name: Identifier<'a>,
-    pub data_type: TypeName,
-    pub nullable: bool,
-    pub default_expr: Option<Box<Expr<'a>>>,
-    pub comment: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum AlterDatabaseAction<'a> {
-    RenameDatabase { new_db: Identifier<'a> },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum AlterTableAction<'a> {
-    RenameTable { new_table: Identifier<'a> },
-    AlterTableClusterKey { cluster_by: Vec<Expr<'a>> },
-    DropTableClusterKey,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct CreateStageStmt {
-    pub if_not_exists: bool,
-    pub stage_name: String,
-
-    pub location: String,
-    pub credential_options: BTreeMap<String, String>,
-    pub encryption_options: BTreeMap<String, String>,
-
-    pub file_format_options: BTreeMap<String, String>,
-    pub on_error: String,
-    pub size_limit: usize,
-    pub validation_mode: String,
-    pub comments: String,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum KillTarget {
-    Query,
-    Connection,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum InsertSource<'a> {
-    Streaming {
-        format: String,
-        rest_tokens: &'a [Token<'a>],
-    },
-    Values {
-        rest_tokens: &'a [Token<'a>],
-    },
-    Select {
-        query: Box<Query<'a>>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct CreateUserStmt {
-    pub if_not_exists: bool,
-    pub user: UserIdentity,
-    pub auth_option: AuthOption,
-    pub role_options: Vec<RoleOption>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct AlterUserStmt {
-    // None means current user
-    pub user: Option<UserIdentity>,
-    // None means no change to make
-    pub auth_option: Option<AuthOption>,
-    pub role_options: Vec<RoleOption>,
-}
-
-#[derive(Debug, Clone, PartialEq, Default)]
-pub struct AuthOption {
-    pub auth_type: Option<AuthType>,
-    pub password: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum RoleOption {
-    TenantSetting,
-    NoTenantSetting,
-    ConfigReload,
-    NoConfigReload,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct AccountMgrStatement {
-    pub source: AccountMgrSource,
-    pub principal: PrincipalIdentity,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum AccountMgrSource {
-    Role {
-        role: String,
-    },
-    Privs {
-        privileges: Vec<UserPrivilegeType>,
-        level: AccountMgrLevel,
-    },
-    ALL {
-        level: AccountMgrLevel,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum AccountMgrLevel {
-    Global,
-    Database(Option<String>),
-    Table(Option<String>, String),
-}
-
-impl RoleOption {
-    pub fn apply(&self, option: &mut UserOption) {
-        match self {
-            Self::TenantSetting => {
-                option.set_option_flag(UserOptionFlag::TenantSetting);
-            }
-            Self::NoTenantSetting => {
-                option.unset_option_flag(UserOptionFlag::TenantSetting);
-            }
-            Self::ConfigReload => {
-                option.set_option_flag(UserOptionFlag::ConfigReload);
-            }
-            Self::NoConfigReload => {
-                option.unset_option_flag(UserOptionFlag::ConfigReload);
-            }
-        }
-    }
-}
-
-impl<'a> Display for ShowLimit<'a> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ShowLimit::Like { pattern } => write!(f, "LIKE {pattern}"),
-            ShowLimit::Where { selection } => write!(f, "WHERE {selection}"),
-        }
-    }
-}
-
-impl<'a> Display for ColumnDefinition<'a> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{} {}", self.name, self.data_type)?;
-        if self.nullable {
-            write!(f, " NULL")?;
-        } else {
-            write!(f, " NOT NULL")?;
-        }
-        if let Some(default_expr) = &self.default_expr {
-            write!(f, " DEFAULT {default_expr}")?;
-        }
-        if let Some(comment) = &self.comment {
-            write!(f, " COMMENT '{comment}'")?;
-        }
-        Ok(())
-    }
-}
-
-impl Display for OptimizeTableAction {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            OptimizeTableAction::All => write!(f, "ALL"),
-            OptimizeTableAction::Purge => write!(f, "PURGE"),
-            OptimizeTableAction::Compact => write!(f, "COMPACT"),
-        }
-    }
-}
-
-impl Display for TableOption {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            TableOption::Engine(engine) => write!(f, "ENGINE = {engine}"),
-            TableOption::Comment(comment) => write!(f, "COMMENT = {comment}"),
-        }
-    }
-}
-
-impl Display for AccountMgrSource {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            AccountMgrSource::Role { role } => write!(f, " ROLE {role}")?,
-            AccountMgrSource::Privs { privileges, level } => {
-                write!(f, " ")?;
-                write_comma_separated_list(f, privileges.iter().map(|p| p.to_string()))?;
-                write!(f, " ON")?;
-                match level {
-                    AccountMgrLevel::Global => write!(f, " *.*")?,
-                    AccountMgrLevel::Database(database_name) => {
-                        if let Some(database_name) = database_name {
-                            write!(f, " {database_name}.*")?;
-                        } else {
-                            write!(f, " *")?;
-                        }
-                    }
-                    AccountMgrLevel::Table(database_name, table_name) => {
-                        if let Some(database_name) = database_name {
-                            write!(f, " {database_name}.{table_name}")?;
-                        } else {
-                            write!(f, " {table_name}")?;
-                        }
-                    }
-                }
-            }
-            AccountMgrSource::ALL { level, .. } => {
-                write!(f, " ALL PRIVILEGES")?;
-                write!(f, " ON")?;
-                match level {
-                    AccountMgrLevel::Global => write!(f, " *.*")?,
-                    AccountMgrLevel::Database(database_name) => {
-                        if let Some(database_name) = database_name {
-                            write!(f, " {database_name}.*")?;
-                        } else {
-                            write!(f, " *")?;
-                        }
-                    }
-                    AccountMgrLevel::Table(database_name, table_name) => {
-                        if let Some(database_name) = database_name {
-                            write!(f, " {database_name}.{table_name}")?;
-                        } else {
-                            write!(f, " {table_name}")?;
-                        }
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-}
-
-impl Display for DatabaseEngine {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        if let DatabaseEngine::Github(token) = self {
-            write!(f, "GITHUB(token=\'{token}\')")
-        } else {
-            write!(f, "DEFAULT")
-        }
-    }
-}
-
-impl Display for Engine {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Engine::Null => write!(f, "NULL"),
-            Engine::Memory => write!(f, "MEMORY"),
-            Engine::Fuse => write!(f, "FUSE"),
-            Engine::Github => write!(f, "GITHUB"),
-            Engine::View => write!(f, "VIEW"),
-            Engine::Random => write!(f, "RANDOM"),
-        }
-    }
-}
-
-impl Display for KillTarget {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            KillTarget::Query => write!(f, "QUERY"),
-            KillTarget::Connection => write!(f, "CONNECTION"),
-        }
-    }
-}
-
-impl Display for CopyUnit<'_> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            CopyUnit::Table {
-                catalog,
-                database,
-                table,
-            } => {
-                if let Some(catalog) = catalog {
-                    write!(
-                        f,
-                        "{catalog}.{}.{table}",
-                        database.as_ref().expect("database must be valid")
-                    )
-                } else if let Some(database) = database {
-                    write!(f, "{database}.{table}")
-                } else {
-                    write!(f, "{table}")
-                }
-            }
-            CopyUnit::StageLocation { name, path } => {
-                write!(f, "@{name}{path}")
-            }
-            CopyUnit::UriLocation {
-                protocol,
-                name,
-                path,
-                credentials,
-                encryption,
-            } => {
-                write!(f, "'{protocol}://{name}{path}'")?;
-                if !credentials.is_empty() {
-                    write!(f, " CREDENTIALS = ( ")?;
-                    write_space_seperated_map(f, credentials)?;
-                    write!(f, " )")?;
-                }
-                if !encryption.is_empty() {
-                    write!(f, " ENCRYPTION = ( ")?;
-                    write_space_seperated_map(f, encryption)?;
-                    write!(f, " )")?;
-                }
-                Ok(())
-            }
-            CopyUnit::Query(query) => {
-                write!(f, "({query})")
-            }
-        }
-    }
-}
-
-impl Display for RoleOption {
-    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-        match self {
-            RoleOption::TenantSetting => write!(f, "TENANTSETTING"),
-            RoleOption::NoTenantSetting => write!(f, "NOTENANTSETTING"),
-            RoleOption::ConfigReload => write!(f, "CONFIGRELOAD"),
-            RoleOption::NoConfigReload => write!(f, "NOCONFIGRELOAD"),
-        }
-    }
 }
 
 impl<'a> Display for Statement<'a> {
